requirements.txt
flask==2.0.1
requests==2.26.0
numpy==1.21.2
python-dotenv==0.19.0